requirements.txt
six
tqdm
torchtext==0.1.1
future
